% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_your_own_pesticide.R
\name{add_your_own_pesticide}
\alias{add_your_own_pesticide}
\title{Add New Pesticides to the PRM Calculation Information Table}
\usage{
add_your_own_pesticide(
  analytes,
  relative_LORs,
  Pesticide.types,
  Distribution.types,
  log.alphas = NULL,
  alphas = NULL,
  log.betas = NULL,
  betas = NULL,
  log.bs = NULL,
  bs = NULL,
  log.cs = NULL,
  cs = NULL,
  log.ks = NULL,
  ks = NULL,
  local.log.alphas = NULL,
  local.alphas = NULL,
  local.log.betas = NULL,
  local.betas = NULL
)
}
\arguments{
\item{local.betas}{}
}
\description{
Add New Pesticides to the PRM Calculation Information Table
}
