% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_your_own_pollutant.R
\name{add_your_own_pollutant}
\alias{add_your_own_pollutant}
\title{Add New Pollutants to the PRM Calculation Information Table}
\usage{
add_your_own_pollutant(
  pollutants,
  relative_LORs,
  pollutant_types,
  distribution_types,
  shape_locations = NA,
  shape_location_2s = NA,
  scales = NA,
  scale_2s = NA,
  weights = NA,
  pollutant_info = CatchThemAll.PRM::pollutant_info
)
}
\arguments{
\item{pollutants}{A vector of pollutant names}

\item{relative_LORs}{A vector of relative limit of reporting
replacement values}

\item{pollutant_types}{A vector of the new pollutant's types}

\item{distribution_types}{A vector of the new pollutant's species
sensitive distribution types}

\item{shape_locations}{A vector of shape/location values (if applicable, else put NA)}

\item{shape_location_2s}{A vector of secondary shape/location values (if applicable, else put NA)}

\item{scales}{A vector of scale values (if applicable, else put NA)}

\item{scale_2s}{A vector of secondary scale values (if applicable, else put NA)}

\item{weights}{A vector of weight values (if applicable, else put NA)}

\item{pollutant_info}{A dataset to add pollutants too}
}
\value{
A data frame
}
\description{
Add New Pollutants to the PRM Calculation Information Table
}
\examples{
new <- add_your_own_pollutant(pollutants = "Poison", relative_LORs = 0.023,
pollutant_types = "Poison", distribution_types = "Log-Normal", scales = 0.09,
shape_locations = 0.014)
multiple_new <- add_your_own_pollutant(pollutants = c("Poison", "Acid", "Sludge"),
relative_LORs = c(0.03, 0.01, 0.5), pollutant_types = c("Ghost", "Bug", "Poison"),
distribution_types = c("Log-Normal", "Log-Logistic Log-Logistic", "Burr Type III"),
scales = c(0.3, 0.002, 2),
scale_2s = c(NA, 0.04, NA),  shape_locations = c(1, 0.07, 3),
shape_location_2s = c(NA, 0.14, 2.3), weights = c(NA, 0.08, NA))
}
