% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Burr_Type_III_Formula.R
\name{Burr_Type_III_Formula}
\alias{Burr_Type_III_Formula}
\title{Burr Type III (Inverse Burr) Equation Formula}
\usage{
Burr_Type_III_Formula(analyte_column, parameter_b, parameter_c, parameter_k)
}
\arguments{
\item{analyte_column}{The column of concentration values for a selected pesticide,
that has a species sensitivity distribution that fits the Burr Type III distribution,
to run the equation on.}

\item{parameter_b}{The Burr Type III b value for the selected pesticide.
These can be found in the "Pesticide_Info" data frame provided in this package.
If you are including other pesticides you will need to append them with their
respective distribution variables to the "Pesticide_Info" table.}

\item{parameter_c}{The Burr Type III c value for the selected pesticide.
These can be found in the "Pesticide_Info" data frame provided in this package.
If you are including other pesticides you will need to append them
with their respective distribution variables to the "Pesticide_Info" table.}

\item{parameter_k}{The Burr Type III k value for the selected pesticide.
These can be found in the "Pesticide_Info" data frame provided in this package.
If you are including other pesticides you will need to append them
with their respective distribution variables to the "Pesticide_Info" table.}
}
\value{
a numeric vector
}
\description{
Burr Type III (Inverse Burr) Equation Formula
}
\examples{
#Chlorpyrifos is used as its species sensitivity distribution plot fits Burr Type III
Chlorpyrifos <- c(0.000000001, 0.5, 2.7, 11)
Burr_Type_III_Formula(analyte_column = Chlorpyrifos,
parameter_b = Pesticide_Info$b[Pesticide_Info$analyte == "Chlorpyrifos"],
parameter_c = Pesticide_Info$c[Pesticide_Info$analyte == "Chlorpyrifos"],
parameter_k = Pesticide_Info$k[Pesticide_Info$analyte == "Chlorpyrifos"])
}
