% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Canto_pesticides}
\alias{Canto_pesticides}
\title{Canto Region Pesticide Concentration Values (Example Data Set)}
\format{
\subsection{\code{Canto_pesticides}}{

A data frame with 808 rows and 64 columns:
\describe{
\item{Site Name}{Site name}
\item{Date}{Sampling date}
\item{Ametryn, Atrazine, Chlorpyrifos, Diuron, Fipronil, Fluroxypyr, Haloxyfop (acid), Hexazinone,
Imazapic, Imidacloprid, Isoxaflutole metabolite (DKN), MCPA, Metolachlor, Metribuzin, Metsulfuron methyl,
Pendimethalin, Prometryn, Simazine, Tebuthiuron, Terbuthylazine, Triclopyr, 2,4-D, Bromacil
Diazinon, Imidicloprid1:40}{pesticide concentration values in ug/L}
...
}
}
}
\usage{
Canto_pesticides
}
\description{
A subset of pesticide concentration data for all pesticides in "pesticide_info" created for this package with fabricated sites
}
\keyword{datasets}
