% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_Sampling_Year.R
\name{find_Sampling_Year}
\alias{find_Sampling_Year}
\title{Title}
\usage{
find_Sampling_Year(dates, wet_season_split = 7)
}
\arguments{
\item{dates}{A date vector from which Sampling Year will be determined. Must be in yyyy-mm-dd format.}

\item{wet_season_split}{The first month of the sampling year in numeric e.g. July = 7}
}
\value{
a factored character vector
}
\description{
Title
}
\examples{
dates <- as.Date("2014-03-04", "2014-12-30", "2015-06-12")
find_Sampling_Year <- (dates) #cut of date for the sampling year will be last day of June
}
