% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_wet_season_end.R
\name{find_wet_season_end}
\alias{find_wet_season_end}
\title{Find Wet Season End Date}
\usage{
find_wet_season_end(wet_season_start, wet_season_length = 182)
}
\arguments{
\item{wet_season_start}{A vector of dates signifying the first day
of the wet season for site year combinations.}

\item{wet_season_length}{The length of the wet season in days.}
}
\value{
A character vector
}
\description{
Find Wet Season End Date
}
\examples{
wet_season_start_dates <- as.Date(c("2014-10-04", "2014-12-30", "2015-09-12"))
wet_season_end_dates <- find_wet_season_end(wet_season_start_dates)
#cut of date for the sampling year will be last day of June
}
