% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_kernal.R
\name{imputation_kernel}
\alias{imputation_kernel}
\title{Imputation Function - Kernal Version}
\usage{
imputation_kernel(impute_variable, wet_season_length = 182)
}
\arguments{
\item{impute_variable}{The variable you wish to impute.}

\item{wet_season_length}{The length of the wet season in days.}
}
\value{
A data frame.
}
\description{
Imputation Function - Kernal Version
}
