% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Kanto_pesticides}
\alias{Kanto_pesticides}
\title{Kanto Region Pesticide Concentration Values (Example Data Set)}
\format{
\subsection{\code{Kanto_pesticides}}{

A data frame with 808 rows and 26 columns:
\describe{
\item{Site Name}{Site name}
\item{Date}{Sampling date}
\item{Ametryn, Atrazine, Chlorpyrifos, Diuron, Fipronil, Fluroxypyr, Haloxyfop (acid), Hexazinone,
Imazapic, Imidacloprid, Isoxaflutole metabolite (DKN), MCPA, Metolachlor, Metribuzin, Metsulfuron methyl,
Pendimethalin, Prometryn, Simazine, Tebuthiuron, Terbuthylazine, Triclopyr, 2,4-D, Bromacil
Diazinon}{Pesticide concentration values in ug/L}
...
}
}
}
\source{
\url{https://bulbapedia.bulbagarden.net/wiki/Kanto}
}
\usage{
Kanto_pesticides
}
\description{
A subset of pesticide concentration data for all pesticides in "Pesticide_Info" from the Kanto region
}
\keyword{datasets}
