% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Log_Logistic_Log_Logistic_Formula.R
\name{Log_Logistic_Log_Logistic_Formula}
\alias{Log_Logistic_Log_Logistic_Formula}
\title{Log Logistic Log Logistic (double curve) CDF Formula}
\usage{
Log_Logistic_Log_Logistic_Formula(
  concentration,
  scale,
  shape_location,
  scale_2,
  shape_location_2,
  weight
)
}
\arguments{
\item{concentration}{The vector of concentration values for a selected pollutant,
that has a species sensitivity distribution that fits the Log Logistic distribution,
to run the equation on.}

\item{scale}{The alpha/scale value for the first log logistic equation for the selected pollutant.
These can be found in the "pollutant_info" data frame provided in this package.
If you are including other pollutants you will need to append them with their
respective distribution variables to the "pollutant_info" table.}

\item{shape_location}{The beta/shape value for the first log logistic equation for the selected pollutant.
These can be found in the "pollutant_info" data frame provided in this package.
If you are including other pollutants you will need to append them with their
respective distribution variables to the "pollutant_info" table.}

\item{scale_2}{The alpha/scale value for the second log logistic equation for the selected pollutant.
These can be found in the "pollutant_info" data frame provided in this package.
If you are including other pollutants you will need to append them with their
respective distribution variables to the "pollutant_info" table.}

\item{shape_location_2}{The beta/shape value for the second log logistic equation for the selected pollutant.
These can be found in the "pollutant_info" data frame provided in this package.
If you are including other pollutants you will need to append them with their
respective distribution variables to the "pollutant_info" table.}

\item{weight}{The weight parameter for combining the two equations for the selected pollutant.
These can be found in the "pollutant_info" data frame provided in this package.
If you are including other pollutants you will need to append them with their
respective distribution variables to the "pollutant_info" table.}
}
\value{
a numeric vector
}
\description{
Log Logistic Log Logistic (double curve) CDF Formula
}
\examples{
twenty3 <- c(0.000000001, 0.5, 2.7, 11)
#Imid 23 is used as its species sensitivity distribution plots fits Log Logistic Log Logistic
Log_Logistic_Log_Logistic_Formula(concentration = twenty3,
scale = pollutant_info$scale[pollutant_info$pollutant == "23"],
shape_location = pollutant_info$shape_location[pollutant_info$pollutant == "23"],
scale_2 = pollutant_info$scale_2[pollutant_info$pollutant == "23"],
shape_location_2 = pollutant_info$shape_location_2[pollutant_info$pollutant == "23"],
weight = pollutant_info$weight[pollutant_info$pollutant == "23"])

}
