% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Log_Normal_Formula.R
\name{Log_Normal_Formula}
\alias{Log_Normal_Formula}
\title{Log-Normal CDF Equation Formula}
\usage{
Log_Normal_Formula(concentration, shape_location, scale)
}
\arguments{
\item{concentration}{The vector of concentration values for a selected pollutant,
that has a species sensitivity distribution that fits a log normal log normal distribution,
to run the equation on.}

\item{shape_location}{The mu/location value for the log normal equation for the selected pollutant.
These can be found in the "pollutant_info" data frame provided in this package.
If you are including other pollutants you will need to append them with their
respective distribution variables to the "pollutant_info" table.}

\item{scale}{The sigma/scale value for the log normal equation for the selected pollutant.
These can be found in the "pollutant_info" data frame provided in this package.
If you are including other pollutants you will need to append them with their
respective distribution variables to the "pollutant_info" table.}
}
\value{
a numeric vector
}
\description{
Log-Normal CDF Equation Formula
}
\examples{
random <- c(0.000000001, 0.5, 2.7, 11)
Log_Normal_Formula(concentration = random,
shape_location = 0.23,
scale = 1.3)
}
