% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Pesticide_Info}
\alias{Pesticide_Info}
\title{Analyte Information for Pesticide Risk Metric Calculations (Reference Table)}
\format{
\subsection{\code{Pesticide_Info}}{

A data frame with 24 rows and 18 columns:
\describe{
\item{analyte}{Pesticide name}
\item{relative_LOR}{The relative Limit of Reporting replacement value}
\item{Pesticide.Type}{The pesticide method of effect}
\item{Distribution.Type}{The species sensetivity distribution shape/type}
\item{log.aplha, alpha, log.beta, beta, log.b, b, log.c, c, log.k, k, local.log.alpha, local.alpha, local.log.beta,
local.beta}{Species sensativity distribution variables used in PRM calculations}
...
}
}
}
\source{
\url{https://www.publications.qld.gov.au/dataset/method-development-pesticide-risk-metric-baseline-condition-of-waterways-to-gbr/resource/c65858f9-d7ba-4aef-aa4f-e148f950220f}
}
\usage{
Pesticide_Info
}
\description{
A reference table for PRM calculations in this package that include Pesticide information
such as type, species sensitivity distribution distributions and relevant equation variables.
}
\keyword{datasets}
