% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wet_season_window.R
\name{plot_wet_season_window}
\alias{plot_wet_season_window}
\title{Plot Wet Season Window Box on plot_daily_PRM}
\usage{
plot_wet_season_window(wet_season_start = 0, wet_season_length = 182)
}
\arguments{
\item{wet_season_start}{The date of the start of the wet season for this
site sampling year.}

\item{wet_season_length}{The length of the wet season in days.}
}
\value{
A plotly shape
}
\description{
Plot Wet Season Window Box on plot_daily_PRM
}
\examples{
shape <- plot_wet_season_window(wet_season_start = "2017-08-01")
}
