% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pollutant_info}
\alias{pollutant_info}
\title{Pollutant Information for Pollutant Risk Metric Calculations (Reference Table)}
\format{
\subsection{\code{pollutant_info}}{

A data frame with 62 rows and 9 columns:
\describe{
\item{pollutant}{Pollutant name}
\item{relative_LOR}{The relative Limit of Reporting replacement value}
\item{pollutant_type}{The pollutant method of effect}
\item{distribution_type}{The species sensetivity distribution shape/type}
\item{shape_location, shape_location2, scale, scale2,
weight}{Species sensativity distribution variables used in PRM calculations}
...
}
}
}
\source{
\url{https://www.publications.qld.gov.au/dataset/method-development-pesticide-risk-metric-baseline-condition-of-waterways-to-gbr/resource/c65858f9-d7ba-4aef-aa4f-e148f950220f}
}
\usage{
pollutant_info
}
\description{
A reference table for PRM calculations in this package that include pollutant information
such as type, species sensitivity distribution distributions and relevant equation variables.
}
\keyword{datasets}
