% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_your_own_pesticide.R
\name{add_your_own_pesticide}
\alias{add_your_own_pesticide}
\title{Add New Pollutants to the PRM Calculation Information Table}
\usage{
add_your_own_pesticide(
  analytes,
  relative_LORs,
  Pesticide.types,
  Distribution.types,
  shape_location = NA,
  shape_location_2 = NA,
  scale = NA,
  scale_2 = NA,
  weight = NA
)
}
\arguments{
\item{analytes}{A vector of pollutant names}

\item{relative_LORs}{A vector of relative limit of reporting
replacement values}

\item{Pesticide.types}{A vector of the new pesticide's types}

\item{Distribution.types}{A vector of the new pesticide's species
sensitive distribution types}

\item{shape_location}{A vector of shape/location values (if applicable, else put NA)}

\item{shape_location_2}{A vector of secondary shape/location values (if applicable, else put NA)}

\item{scale}{A vector of scale values (if applicable, else put NA)}

\item{scale_2}{A vector of secondary scale values (if applicable, else put NA)}

\item{weight}{A vector of weight values (if applicable, else put NA)}
}
\value{
A data frame
}
\description{
Add New Pollutants to the PRM Calculation Information Table
}
\examples{
new <- add_your_own_pesticide(analytes = "Poison", relative_LORs = 0.023,
Pesticide.types = "Poison", Distribution.types = "Log-Normal", scale = 0.09,
shape_location = 0.014)
multiple_new <- add_your_own_pesticide(analytes = c("Poison", "Acid", "Sludge"),
relative_LORs = c(0.03, 0.01, 0.5), Pesticide.types = c("Ghost", "Bug", "Poison"),
Distribution.types = c("Log-Normal", "Log-Logistic Log-Logistic", "Burr Type III"),
scale = c(0.3, 0.002, 2),
scale_2 = c(NA, 0.04, NA),  shape_location = c(1, 0.07, 3),
shape_location_2 = c(NA, 0.14, 2.3), weight = c(NA, 0.08, NA))
}
