% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gamma_Formula.R
\name{Gamma_Formula}
\alias{Gamma_Formula}
\title{Gamma CDF Equation Formula}
\usage{
Gamma_Formula(concentration, shape_location, scale)
}
\arguments{
\item{concentration}{The vector of concentration values for a selected pollutant,
that has a species sensitivity distribution that fits a gamma distribution,
to run the equation on.}

\item{shape_location}{The k/shape value for the gamma equation for the selected pollutant.
These can be found in the "pollutant_info" data frame provided in this package.
If you are including other pollutants you will need to append them with their
respective distribution variables to the "pollutant_info" table.}

\item{scale}{The scale/theta value for the gamma equation for the selected pollutant.
These can be found in the "pollutant_info" data frame provided in this package.
If you are including other pollutants you will need to append them with their
respective distribution variables to the "pollutant_info" table.}
}
\value{
a numeric vector
}
\description{
Gamma CDF Equation Formula
}
\examples{
#Imidacloprid is used as its species sensitivity distribution plot fits log-normal log-normal
Imidacloprid <- c(0.000000001, 0.5, 2.7, 11)
Gamma_Formula(concentration = Imidacloprid,
shape_location = pollutant_info$shape_location[pollutant_info$pollutant == "Imidacloprid"],
scale = pollutant_info$scale[pollutant_info$pollutant == "Imidacloprid"])
}
