% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Log_Gumbel_Formula.R
\name{Log_Gumbel_Formula}
\alias{Log_Gumbel_Formula}
\title{Log Gumbel CDF Equation Formula}
\usage{
Log_Gumbel_Formula(concentration, shape_location, scale)
}
\arguments{
\item{concentration}{The vector of concentration values for a selected pesticide,
that has a species sensitivity distribution that fits a log gumbel distribution,
to run the equation on.}

\item{shape_location}{The mu/location value for the log gumbel equation for the selected pesticide.
These can be found in the "Pesticide_Info" data frame provided in this package.
If you are including other pesticides you will need to append them with their
respective distribution variables to the "Pesticide_Info" table.}

\item{scale}{The beta/scale value for the log gumbel equation for the selected pesticide.
These can be found in the "Pesticide_Info" data frame provided in this package.
If you are including other pesticides you will need to append them with their
respective distribution variables to the "Pesticide_Info" table.}
}
\value{
a numeric vector
}
\description{
Log Gumbel CDF Equation Formula
}
\examples{
#Imidacloprid is used as its species sensitivity distribution plot fits log-normal log-normal
Imidacloprid <- c(0.000000001, 0.5, 2.7, 11)
Log_Gumbel_Formula(concentration = Imidacloprid,
shape_location = Pesticide_Info$shape_location[Pesticide_Info$analyte == "Imidacloprid"],
scale = Pesticide_Info$scale[Pesticide_Info$analyte == "Imidacloprid"])
}
