% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Log_Normal_Log_Normal_Formula.R
\name{Log_Normal_Log_Normal_Formula}
\alias{Log_Normal_Log_Normal_Formula}
\title{Log-Normal Log-Normal (double curve) CDF Equation Formula}
\usage{
Log_Normal_Log_Normal_Formula(
  concentration,
  shape_location,
  scale,
  shape_location_2,
  scale_2,
  weight
)
}
\arguments{
\item{concentration}{The vector of concentration values for a selected pollutant,
that has a species sensitivity distribution that fits a log normal log normal distribution,
to run the equation on.}

\item{shape_location}{The mu/shape value for the first log normal equation for the selected pollutant.
These can be found in the "pollutant_info" data frame provided in this package.
If you are including other pollutants you will need to append them with their
respective distribution variables to the "pollutant_info" table.}

\item{scale}{The sigma/scale value for the first log normal equation for the selected pollutant.
These can be found in the "pollutant_info" data frame provided in this package.
If you are including other pollutants you will need to append them with their
respective distribution variables to the "pollutant_info" table.}

\item{shape_location_2}{The mu/shape value for the second log normal equation for the selected pollutant.
These can be found in the "pollutant_info" data frame provided in this package.
If you are including other pollutants you will need to append them with their
respective distribution variables to the "pollutant_info" table.}

\item{scale_2}{The sigma/scale value for the second log normal equation for the selected pollutant.
These can be found in the "pollutant_info" data frame provided in this package.
If you are including other pollutants you will need to append them with their
respective distribution variables to the "pollutant_info" table.}

\item{weight}{The weight parameter for combining the two equations for the selected pollutant.
These can be found in the "pollutant_info" data frame provided in this package.
If you are including other pollutants you will need to append them with their
respective distribution variables to the "pollutant_info" table.}
}
\value{
a numeric vector
}
\description{
Log-Normal Log-Normal (double curve) CDF Equation Formula
}
\examples{
#Imidacloprid 38 is used as its species sensitivity distribution plot fits log-normal log-normal
Imid38 <- c(0.000000001, 0.5, 2.7, 11)
Log_Normal_Log_Normal_Formula(concentration = Imid38,
shape_location = pollutant_info$shape_location[pollutant_info$pollutant == "38"],
scale = pollutant_info$scale[pollutant_info$pollutant == "38"],
shape_location_2 = pollutant_info$shape_location_2[pollutant_info$pollutant == "38"],
scale_2 = pollutant_info$scale_2[pollutant_info$pollutant == "38"],
weight = pollutant_info$weight[pollutant_info$pollutant == "38"])
}
